+++ src/panels/lovelace/cards/hui-conditional-card.ts
@@ -1,4 +1,4 @@
-import { createCardElement } from "../common/create-card-element";
+import { createCardElement } from "../create-element/create-card-element";
 import { computeCardSize } from "../common/compute-card-size";
 import {
   checkConditionsMet,
