+++ src/panels/lovelace/cards/hui-stack-card.ts
@@ -1,6 +1,6 @@
 import { html, LitElement, TemplateResult, CSSResult, css } from "lit-element";
 
-import { createCardElement } from "../common/create-card-element";
+import { createCardElement } from "../create-element/create-card-element";
 import { LovelaceCard } from "../types";
 import { LovelaceCardConfig } from "../../../data/lovelace";
 import { HomeAssistant } from "../../../types";
