+++ src/panels/lovelace/cards/picture-elements/create-styled-hui-element.ts
@@ -1,5 +1,5 @@
 import { LovelaceElement, LovelaceElementConfig } from "../../elements/types";
-import { createHuiElement } from "../../common/create-hui-element";
+import { createHuiElement } from "../../create-element/create-hui-element";
 
 export function createStyledHuiElement(
   elementConfig: LovelaceElementConfig
