+++ src/panels/lovelace/create-element/create-badge-element.ts
@@ -2,7 +2,7 @@ import "../badges/hui-entity-filter-badge";
 import "../badges/hui-state-label-badge";
 
 import { LovelaceBadgeConfig } from "../../../data/lovelace";
-import { createLovelaceElement } from "../create-element/create-element-base";
+import { createLovelaceElement } from "./create-element-base";
 
 const BADGE_TYPES = new Set(["entity-filter", "error", "state-label"]);
